add invalid uri template test for serviceInfo & cleamup
